Restructured app
