+++ blueprints/screening_app/entities/gender/mapping.json
@@ -21,10 +21,8 @@
             "machito",
             "muchachito",
             "machin"
-         
-]
-      
-},
+         ]
+      },
       {
          "id":"102843",
          "cname":"Mujer",
@@ -47,10 +45,7 @@
             "dama",
             "damita",
             "damicela"
-         
-]
-      
-}
-   
-]
+         ]
+      }
+   ]
 }
