+++ blueprints/screening_app/root.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""This module contains an empty application container.
+It is defined here to avoid circular imports
+"""
+from mindmeld import Application
+
+app = Application(__name__)
+
+__all__ = ['app']
